requirements.txt
appdirs==1.4.4
asgiref==3.4.1
astroid==2.6.6
black==21.7b0
click==8.0.1
Django==3.2.6
django-environ==0.4.5
django-filter==2.4.0
djangorestframework==3.12.4
flake8==3.9.2
isort==5.9.3
lazy-object-proxy==1.6.0
mccabe==0.6.1
mypy-extensions==0.4.3
mysqlclient==2.0.3
pathspec==0.9.0
Pillow==8.3.1
pycodestyle==2.7.0
pyflakes==2.3.1
pylint==2.9.6
pytz==2021.1
regex==2021.8.3
sqlparse==0.4.1
toml==0.10.2
tomli==1.2.1
wrapt==1.12.1
